добавлен начальный уровень спринг секьюрити

+++ src/main/java/ru/igor/PP_3_1_3/dao/UserDao.java
@@ -11,7 +11,7 @@ public interface UserDao {
 
     void updateUser(User updateUser);
 
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
 
-    User getUser(Integer id);
+    User getUser(Long id);
 }
